Add support for CPLEX
The Problem.solve method now accepts CPLEX when using the solver
keyword argument. As expected, the solver output will be hidden by
default, but can be shown by setting the verbose keyword argument to
True. The warm_start solver option is supported for linear
constraints. There are two CPLEX-specific solver options which are
supported: 1) cplex_params is a keyword argument that accepts a
dictionary of parameter number keys and parameter values; 2)
cplex_filename takes a file path which will be used to export the
model to a given file format (e.g., SAV, LP, MPS) before solving.

For example, the following snippet will set the solver to CPLEX,
display the engine log, set a CPLEX parameter, and export the model to
LP format:

>>> # Set CPXPARAM_Benders_Strategy (1501) to FULL (3)
... cpx_params = {1501: 3}
>>> prob.solve(solver=CPLEX, verbose=True, cplex_params=cpx_params,
               cplex_filename="example.lp")