Allow test_examples.py to be run as a standalone script and test CPLEX
Previously, there were two tests named test_advanced, so only the
later would every run. They are now named test_advanced1 and
test_advanced2.

NOTE: The test_log_det fails with SCS (which is the default for my
setup). It would probably be a good idea to mark that with
unittest.skip(), but I'll leave that to the authors.