When gathering dual values, no longer need to check for None
Now that we put empty constraints in the model, rather than using None
as a placeholder, we no longer need to check for None when gathering
dual values. An assert was added to ensure this. A FIXME comment was
also removed about calculating qcp duals; due to the way quadratic
constraints are modeled, the quadratic constraints are never queried
directly.